+++ js/const.js
@@ -1,0 +1,43 @@
+export const POSTS_LENGTH = 25;
+
+export const DESCRIPTIONS = [
+  'photo-1',
+  'photo-2',
+  'photo-3',
+  'photo-4',
+  'photo-5',
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const NAMES = [
+  'Вася',
+  'Алиса',
+  'Горыныч',
+  'Кнопычка',
+  'МариИвановна',
+  'Димас',
+  'Серый',
+  'Ли Джун Ки',
+  'Силвестер',
+  'Скала',
+  'Adele',
+  'Ais',
+  'Ромчик',
+  'Карп',
+  'Луна',
+  'Васелиса',
+  'Анджела',
+  'Грут',
+  'Джонс',
+  'Манки',
+  'Хрюша',
+];
+
